% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amd10y}
\alias{amd10y}
\title{Ten year neovascular AMD survival data}
\format{
A data frame (tibble) with 6696 rows and 23 variables:
\describe{
\item{patID}{Anonymized patient identifier.}
\item{sex}{Sex of patient (m = male, f = female)}
\item{time}{time in days after date of first injection (0 = first injection or baseline)}
\item{eye}{Left or right eye of patient (r = right, l = left)}
\item{ethnicity}{Ethnicity of patient following the categories from the
UK ETHNIC CATEGORY CODE 2001. (asian = Asian or Asian British,
white = White, black = Black or Black British,
other_unknown = Other ethnic group or unknown)}
\item{fellow}{fellow eye requires anti-VEGF within observation period}
\item{ttofellow_mo}{time between baseline and fellow eye involvement in months}
\item{drug_switch}{anti-VEGF drug agent is switched during observation period }
\item{ttodrugswitch_mo}{time between baseline and date of drug switch in months}
\item{ltfu}{patient is lost to follow-up during observation period }
\item{ltfu_stable}{of those lost to follow-up, are they stable prior to this timepoint }
\item{ltfu_outcome}{reason for lost to follow-up (discharged, dna (did not attend last appointment), dead, transferred)}
\item{ttoltfu_mo}{time between baseline and loss to follow-up in months}
\item{ten_y_stable}{patient is stable at ten year timepoint }
\item{irf}{presence of intraretinal fluid at last visit }
\item{srf}{presence of subretinal fluid at last visit }
\item{disciform}{presence of disciform scar at last visit (0 = no, 1a = type 1a, 1b = type 1b)}
\item{ga}{presence of geographic atrophy at last visit  }
\item{ga_foveal}{presence of foveal geographic atrophy at last visit}
\item{ttoga_mo}{time to geographic atrophy development in months}
\item{inj_num}{injection number}
\item{va}{visual acuity in Early Treatment Diabetic Retinopathy Study letters}
\item{crt}{foveal thickness in microns}
}
@source \url{https://doi.org/10.5061/dryad.9cnp5hqfm}
}
\usage{
data("amd10y")
}
\description{
Ten Year Survival Trends of Neovascular Age Related Macular Degeneration
at First Presentation
}
\section{Missing values}{

To be added
}

\section{Study setting and design}{

Retrospective cohort study of treatment-naïve, first-affected eyes with
nAMD started on ranibizumab before January 1, 2009.
Approval was granted by the Institutional Review Board of the hospital
(ROAD17/031). The study complied with the Declaration of Helsinki
}

\section{Data source}{

All clinical information at Moorfields Eye Hospital is recorded within an
electronic medical record (EMR) application (OpenEyes Foundation, London, UK).
A SQL database (SQL Server Reporting Service, Microsoft Corporation, Richmond, USA)
containing all the information from the EMR is in place and regular updates
are performed overnight to keep the data warehouse up-to-date.
VA is reported in ETDRS letter score (and categories for ETDRS < 1).
The highest value (independent of measurement method)
available at each visit was chosen.
}

\section{Data}{

Types of data: de-identified human subjects data
Information governance authorised Moorfields Eye Clinic 19/07/2018.
age not provided as a continuous variable as in original analysis
to facilitate de-identification
}

\seealso{
Scientific article to which this data set was supplement:
\url{https://doi.org/10.1136/bjophthalmol-2020-317161}
}
\keyword{datasets}
