% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{anonymize}
\alias{anonymize}
\alias{anonymise}
\title{anonymize}
\usage{
anonymize(
  df,
  simple_names = TRUE,
  replace_rownames = FALSE,
  colString = "var",
  rowString = "sample"
)

anonymise(
  df,
  simple_names = TRUE,
  replace_rownames = FALSE,
  colString = "var",
  rowString = "sample"
)
}
\arguments{
\item{df}{data frame to anonymise}

\item{simple_names}{logical, column names will be converted to letters. If FALSE, colString will be used for renaming}

\item{replace_rownames}{if TRUE, replacement with string from rowString argument}
}
\value{
Data frame
}
\description{
The function will anonymize your data.
Modified from \href{https://stackoverflow.com/a/10455729/7941188}{this} and \href{https://stackoverflow.com/a/10458688/7941188}{this}
post on Stackoverflow from user Etienne Low-Décarie and Gavin Simpson
}
\seealso{
Other convenience functions: 
\code{\link{csv}()},
\code{\link{get_age}()},
\code{\link{show_stats}()}
}
\concept{convenience functions}
