% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink.R
\name{blink}
\alias{blink}
\title{Your data in a blink of an eye}
\usage{
blink(x, id = NULL, eye = NULL, to = "logmar", ...)
}
\arguments{
\item{x}{data frame}
}
\description{
High level function, wrapping \link{myop}, \link{eyes} and \link{reveal}
}
\details{
Steps of data processing:
\itemize{
\item checking for duplicate rows and removing them
\item Renaming columns to prepare for myopization and VA conversion
\item VA will always be converted to logmar by default
}
}
\examples{
TBC
}
