% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cleaner.R
\name{convert_NLP}
\alias{convert_NLP}
\alias{isNAstring}
\title{VA cleaner helper}
\usage{
convert_NLP(x, replace_PL = c(pl = "lp", npl = "nlp"))

isNAstring(x, full = c("\\\\.+", "", "\\\\s+", "n/a", "na", "null"))
}
\arguments{
\item{replace_PL}{named vector how to rename qualitative VA}

\item{full}{vector of full strings to be replaced by NA}

\item{partial}{vector of partial strings to be replaced by NA}
}
\description{
VA cleaner helper
}
\seealso{
Other VA cleaner: 
\code{\link{va_cleaner}},
\code{\link{va}()}
}
\concept{VA cleaner}
