% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{count_eyes}
\alias{count_eyes}
\title{internal count}
\usage{
count_eyes(x, pat_col, eye)
}
\arguments{
\item{x}{object (data frame)}

\item{pat_col}{patient column}

\item{eye}{eye column}
}
\description{
\code{count_eyes()} is the internal counting function
}
\keyword{[`eyes()`]}
\keyword{for}
\keyword{internal}
