% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{csv}
\alias{csv}
\title{csv}
\usage{
csv(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{data frame}

\item{name}{Filename. Default: Name of dataframe to save as csv. Or character string (.csv extension added automatically)}
}
\description{
wrapper around write.csv with default 'row.names = FALSE'.
Will use the name of the data frame for the generated .csv file.
}
\seealso{
Other convenience functions: 
\code{\link{anonymize}()},
\code{\link{get_age}()},
\code{\link{show_stats}()}
}
\concept{convenience functions}
