% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye-package.R
\docType{package}
\name{eye-package}
\alias{eye-package}
\title{eye: Analysis of Eye Data}
\description{
There is no ophthalmic researcher who has not had headaches from the handling of visual acuity entries. Different notations, untidy entries. This shall now be a matter of the past. Eye makes it as easy as pie to work with VA data - easy cleaning, easy conversion between Snellen, logMAR, ETDRS letters, and qualitative visual acuity shall never pester you again. The eye package automates the pesky task to count number of patients and eyes, and can help to clean data with easy re-coding for right and left eyes. It also contains functions to help reshaping eye side specific variables between wide and long format. Visual acuity conversion is based on Schulze-Bonsel et al. (2006) \doi{10.1167/iovs.05-0981}, Gregori et al. (2010) \doi{10.1097/iae.0b013e3181d87e04}, Beck et al. (2003) \doi{10.1016/s0002-9394(02)01825-1} and Bach (2007) \url{http://michaelbach.de/sci/acuity.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjebo/eye}
  \item Report bugs at \url{https://github.com/tjebo/eye/issues}
}

}
\author{
\strong{Maintainer}: Tjebo Heeren \email{tjebo@gmx.de} (\href{https://orcid.org/0000-0001-5297-2301}{ORCID})

Other contributors:
\itemize{
  \item Antoine Fabri [contributor]
}

}
\keyword{internal}
