% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_helper.R
\docType{data}
\name{eye_codes}
\alias{eye_codes}
\title{List of codes}
\format{
An object of class \code{list} of length 11.
}
\usage{
eye_codes
}
\description{
List of codes
}
\details{
\itemize{
\item r right eyes
\item l left eyes
\item iop IOP codes
\item va VA codes
\item method VA methods
\item id patient column codes
\item quali quali VA codes
}
}
\keyword{internal}
