% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{eyes_to_string}
\alias{eyes_to_string}
\title{Eye count to strings}
\usage{
eyes_to_string(x, english = TRUE, para = FALSE, UK = FALSE)
}
\arguments{
\item{x}{vector of one or two}

\item{english}{If TRUE: writing numbers <= 12 as words}

\item{para}{If TRUE: Adding "A total of" to comply with most
journal standards and to avoid awkward long numnbers.}

\item{UK}{Logical, Use UK (English) style (TRUE) or
USA (American) style (FALSE).}
}
\value{
String to paste into text
}
\description{
Eye count to strings
}
\seealso{
Other Eye count functions: 
\code{\link{eyes}()}
}
\concept{Eye count functions}
