% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trail.R
\docType{data}
\name{geom_trail}
\alias{geom_trail}
\alias{GeomTrail}
\title{geom_trail}
\format{
An object of class \code{GeomTrail} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
geom_trail(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

GeomTrail
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Mark a trail with a plot just like the base plot type = "b".
You can also leave the dots blank, thus allowing use of text instead of
points (see examples).
}
\section{Additional arguments}{

\strong{\code{gap}} gap between points and lines
}

\section{Aesthetics}{

\code{geom_trail} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{color}
\item \code{linetype}
\item \code{size} size of points - 0 for empty space
\item \code{linesize}
}
}

\examples{
library(ggplot2)
library(dplyr)

ggplot(pressure, aes(temperature, pressure)) +
  geom_ribbon(aes(ymin = pressure - 50, ymax = pressure + 50), alpha = 0.2) +
  geom_trail()

amd_aggr <-
amd \%>\%
  group_by(
    age_cut10 = cut_width(BaselineAge, 10),
    days_cut90 = cut_width(FollowupDays, 90, labels = seq(0, 810, 90))
  ) \%>\%
  summarise(mean_va = mean(VA_ETDRS_Letters))

p <- ggplot(amd_aggr, aes(days_cut90, mean_va, color = age_cut10)) +
       theme_classic() +
       labs(
         x = "Follow up time [Days]", y = "Mean VA [ETDRS letters]",
         color = "Age strata"
       )

p + geom_trail(aes(group = age_cut10))

p + geom_trail(aes(group = age_cut10), size = 0) +
  geom_text(aes(label = round(mean_va, 0)), show.legend = FALSE)
}
\seealso{
The geom was modified from the suggestion by user teunbrand on
Stackoverflow in
\href{https://stackoverflow.com/a/55857158/7941188}{this thread}
}
\keyword{datasets}
