% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_age}
\alias{get_age}
\title{get_age}
\usage{
get_age(from_date, to_date = lubridate::now(), period = FALSE)
}
\arguments{
\item{from_date}{start date}

\item{to_date}{end date}

\item{period}{default FALSE: output as a duration. If TRUE, output as a period}
}
\description{
calculates age in years, either as duration or as period
Slight modification of a function \href{https://stackoverflow.com/a/47529507/7941188}{posted on stackoverflow by user Moody_Mudskipper}
}
\seealso{
Other convenience functions: 
\code{\link{anonymize}()},
\code{\link{csv}()},
\code{\link{show_stats}()}
}
\concept{convenience functions}
