% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getage}
\alias{getage}
\title{getage}
\usage{
getage(from_date, to_date = lubridate::now(), period = FALSE, dec = 1)
}
\arguments{
\item{from_date}{start date}

\item{to_date}{end date}

\item{period}{Calculating period (TRUE) or duration (FALSE- default)}

\item{dec}{How many decimals are displayed}
}
\value{
Numeric vector
}
\description{
calculates age in years, as durations or periods
}
\examples{
getage("1984-10-16")

dob <-  c("1984-10-16", "2000-01-01")
test_date <-  as.Date(dob) + c(15000, 20000)
getage(dob, test_date)
}
\seealso{
\href{https://stackoverflow.com/a/47529507/7941188}{OP on stackoverflow}
from which this function was inspired.
\href{https://lubridate.tidyverse.org/articles/lubridate.html#time-intervals}{Read about periods and durations}
}
\author{
Antoine Fabri and Tjebo Heeren
}
\concept{convenience functions}
