% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{insight}
\alias{insight}
\title{insight}
\usage{
insight(x, ...)
}
\arguments{
\item{x}{data frame}

\item{...}{passed to \link{eyes}}
}
\description{
wrapper around eyes, returning text for document
}
\examples{
insight(amd)
}
\seealso{
Other ophthalmic functions: 
\code{\link{eyes}()}
}
\concept{ophthalmic functions}
