% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cleaner.R
\name{introduceNA}
\alias{introduceNA}
\title{introduce NA for implausible VA entries}
\usage{
introduceNA(x, test)
}
\arguments{
\item{x}{vector}

\item{test}{plausibility test}
}
\value{
vector
}
\description{
introduce NA for implausible VA entries
}
\keyword{internal}
