% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trail.R
\name{makeContent.trail}
\alias{makeContent.trail}
\title{makeContent.trail}
\usage{
\method{makeContent}{trail}(x)
}
\arguments{
\item{x}{grob object passed to method}
}
\description{
underlying drawing method for paths in geom_trail
}
\author{
Teun van den Brand
}
