% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myop.R
\name{myop}
\alias{myop}
\alias{myope}
\alias{myopise}
\alias{myopize}
\alias{myopic}
\alias{pivot_myop}
\title{long eye data}
\usage{
myop(x, id = NULL, ...)

myope(x, id = NULL, ...)

myopise(x, id = NULL, ...)

myopize(x, id = NULL, ...)

myopic(x, id = NULL, ...)

pivot_myop(x, old_cols, eye_code = c("r", "l"), ...)
}
\arguments{
\item{x}{vector}

\item{id}{if there are both IOP and VA columns, one needs to
specify if there is more than one ID column.}

\item{...}{arguments passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}
}
\description{
Wrapper around \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} to make your data
long ("myopic")
}
\details{
\code{myop()} will look for columns that have
typical eye codes, i.e. c("r", "re", "od", "right") for right eyes and
c("l", "le", "os", "left") for left eyes.
}
\examples{
set.seed(42)
iop <- data.frame(id = letters[1:11], r = sample(10:20), l = sample(10:20))

myop(iop, values_to = "iop")

# Example to clean a bit messy data frame
iop_va <- data.frame(id = letters[1:11],
                     iop_r = sample(10:20), iop_l = sample(10:20),
                     va_r = sample(40:50), va_l = sample(40:50))
myop(iop_va)

}
\concept{myopizing functions}
