% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myop.R
\name{myop_pivot}
\alias{myop_pivot}
\title{myopizer}
\usage{
myop_pivot(x)
}
\arguments{
\item{x}{object (data frame)}
}
\description{
The actual myopization using \link[tidyr:pivot_longer]{tidyr::pivot_longer}
several columns with partial strings with "eye codes" in their names
}
\seealso{
Other myopizer: 
\code{\link{myop_rename}()},
\code{\link{myopizer}()}
}
\concept{myopizer}
\keyword{internal}
