% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myop.R
\name{myop_rename}
\alias{myop_rename}
\title{Rename}
\usage{
myop_rename(x)
}
\arguments{
\item{x}{object (data frame)}
}
\description{
Rename data names for Myop
}
\seealso{
Other myopizer: 
\code{\link{myop_pivot}()},
\code{\link{myopizer}()}
}
\concept{myopizer}
