% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myop.R
\name{myopizer}
\alias{myopizer}
\title{myopizer}
\usage{
myopizer(x, var = "value")
}
\arguments{
\item{x}{data frame}

\item{var}{Character vector of length 1 specifying the variable if there
is only one column per eye with no further info on the variable
(default "value")}
}
\description{
internal function, checks and prepares data frames for
"myopization":
Removing duplicates, returning myopized data if criteria fulfilled
(No "eye" column, more than one variable column with eye codes
as partial string). Names need to be prepared with
\link{myop_rename} beforehand.
}
\seealso{
Other myopizer: 
\code{\link{myop_pivot}()},
\code{\link{myop_rename}()}
}
\concept{myopizer}
\keyword{internal}
