% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_methods.R
\name{parse_snellen}
\alias{parse_snellen}
\title{parsing snellen fractions to numeric values}
\usage{
parse_snellen(y)
}
\arguments{
\item{y}{vector}
}
\description{
parsing snellen fractions to numeric values
}
\keyword{internals}
