% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_blink}
\alias{print_blink}
\alias{print.blink}
\title{print blink}
\usage{
\method{print}{blink}(x, ...)
}
\arguments{
\item{x}{object of class "blink"}

\item{...}{arguments passed to \link{print.default}}
}
\description{
S3 methods for blink class
}
\seealso{
Other print methods: 
\code{\link{print_VA}}
}
\concept{print methods}
