% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_VA}
\alias{print_VA}
\alias{print.snellen}
\alias{print.logmar}
\alias{print.etdrs}
\title{print VA class}
\usage{
\method{print}{snellen}(x, ...)

\method{print}{logmar}(x, ...)

\method{print}{etdrs}(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{arguments passed to \link{print.default}}
}
\description{
S3 methods for VA classes "snellen", "logmar" and "etdrs".
\strong{snellen} is always also a character class- because it is more categorical
than continuous. \strong{logmar} and \strong{etdrs} are both numerics
(logMAR is double, etdrs is integer).
}
\seealso{
Other print methods: 
\code{\link{print_blink}}
}
\concept{print methods}
