% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prob_contour}
\alias{prob_contour}
\title{Probability contours}
\usage{
prob_contour(data, x = NULL, y = NULL, n = 50, prob = 0.95, ...)
}
\arguments{
\item{data}{data frame x and y coordinates}

\item{x}{column with x coordinates (default first column)}

\item{y}{column with y coordinates (default second column)}

\item{n}{passed to \link[MASS:kde2d]{MASS::kde2d}}

\item{prob}{probability to be estimated}

\item{...}{further parameters passed to \link[MASS:kde2d]{MASS::kde2d}}
}
\description{
calculates 2d probability contours for use in ggplot2.
Modified from \href{https://stackoverflow.com/a/59173290/7941188}{this post} by user crsh
}
\examples{

library(ggplot2)
set.seed(1)
n=100
foo <- data.frame(x=rnorm(n, 0, 1), y=rnorm(n, 0, 1))

df_contours <- dplyr::bind_rows(
  purrr::map(seq(0.2, 0.8, 0.2), function(p) prob_contour(foo, prob = p))
)

ggplot() +
  geom_point(data = foo, aes(x = x, y = y)) +
  geom_polygon(data = df_contours, aes(x = x, y = y, color = prob), fill = NA) +
  scale_color_brewer(name = "Probs", palette = "Set1")
}
\seealso{
Other stats functions: 
\code{\link{see}()}
}
\concept{stats functions}
