% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{reveal_methods}
\alias{reveal_methods}
\alias{revealEye}
\alias{revealEye.list}
\alias{revealEye.numeric}
\alias{revealEye.data.frame}
\alias{revealEye.default}
\title{reveals little helper}
\usage{
revealEye(x, ...)

\method{revealEye}{list}(x, by, dec, ...)

\method{revealEye}{numeric}(x, dec, ...)

\method{revealEye}{data.frame}(x, dec, ...)

\method{revealEye}{default}(x, dec, ...)
}
\arguments{
\item{x}{atomic vector, list of numeric vectors or data frame}

\item{...}{further arguments passed to methods}

\item{by}{grouping variable - character vector with column names
can be several variables!}

\item{dec}{how many decimals are displayed}
}
\value{
data frame
}
\description{
S3 generic and methods
}
\seealso{
Other revealer: 
\code{\link{reveal_split}()},
\code{\link{reveal}()}
}
\concept{revealer}
