% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{reveal_split}
\alias{reveal_split}
\title{reveals little helper}
\usage{
reveal_split(x, by)
}
\arguments{
\item{x}{data frame}

\item{by}{character vector with the names of the columns}
}
\description{
splits into groups
}
\seealso{
Other revealer: 
\code{\link{reveal_methods}},
\code{\link{reveal}()}
}
\concept{revealer}
\keyword{internal}
