% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_helper.R
\name{set_codes}
\alias{set_codes}
\title{Set string codes}
\usage{
set_codes(
  r = c("r", "re", "od", "right"),
  l = c("l", "le", "os", "left"),
  iop = c("iop", "gat", "nct"),
  va = c("va", "bcva"),
  method = c("etdrs", "snellen", "logmar"),
  id = c("pat", "id"),
  quali = c("nlp", "lp", "hm", "cf"),
  vaPart = "acuit",
  ...
)
}
\arguments{
\item{r}{right eyes}

\item{l}{left eyes}

\item{iop}{IOP codes}

\item{va}{VA codes}

\item{method}{VA methods}

\item{id}{patient column codes}

\item{quali}{quali VA codes}

\item{vaPart}{partial VA codes}

\item{...}{further codes to set}
}
\description{
set the eye strings looked for in the data
}
\seealso{
Other string matching functions: 
\code{\link{getElem}},
\code{\link{sort_substr}()},
\code{\link{str_func_facs}}
}
\concept{string matching functions}
