% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{show_stats}
\alias{show_stats}
\title{show_stats}
\usage{
show_stats(x, dec = 1, rownames = TRUE)
}
\arguments{
\item{x}{either vector or list of vectors. Data frames supported.
Character vectors (or character columns) will be removed.}

\item{dec}{how many decimals are displayed}

\item{rownames}{if FALSE, column with variable names will be created}
}
\value{
named vector (for vector) or data frame (for list)
}
\description{
Shows commonly used summary statistics
}
\examples{
x = y = z = c(rnorm(20), NA)

# named or unnamed list
mylist <- list(x = x, y = y, z = z)
show_stats(mylist)
# with a data frame
mydf <- data.frame(x, y, z)
show_stats(mydf)
#If aggregation by group, split the data frame first
mydf2 <- data.frame(group = rep(letters[1:2], each = 42), x, y, z)
lapply(split(mydf2, mydf2$group), show_stats, rownames = FALSE)
}
\seealso{
Other convenience functions: 
\code{\link{anonymize}()},
\code{\link{csv}()},
\code{\link{get_age}()}
}
\concept{convenience functions}
