% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_helper.R
\name{sort_substr}
\alias{sort_substr}
\title{sort substrings}
\usage{
sort_substr(x, list_substr)
}
\arguments{
\item{x}{vector of strings}

\item{list_substr}{list of substrings to match against and sort -
the order in the list defines the resulting order in the string
Should be names list - the names will be the codes to which the
tokens will be matched against.}
}
\description{
unify code for substrings and arrange, after tokenizing
}
\seealso{
Other string matching functions: 
\code{\link{getElem}},
\code{\link{recodeye}()},
\code{\link{set_codes}()},
\code{\link{str_func_facs}}
}
\concept{string matching functions}
\keyword{internal}
