% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_helper.R
\name{str_func_facs}
\alias{str_func_facs}
\alias{whole_str}
\alias{part_str}
\alias{both_str}
\title{String search func facs}
\usage{
whole_str(string)

part_str(string)

both_str(string_vec)
}
\arguments{
\item{string}{vector of strings to look for}

\item{string_vec}{vector of two strings}
}
\description{
internal function factory for functions
to match "whole_str string" with any non character as boundaries
accepts string as regular expression.
}
\seealso{
Other string matching functions: 
\code{\link{getElem}},
\code{\link{recodeye}()},
\code{\link{set_codes}()},
\code{\link{sort_substr}()}
}
\concept{string matching functions}
\keyword{internal}
