% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_helper.R
\name{str_search}
\alias{str_search}
\alias{whole_str}
\alias{part_str}
\alias{both_str}
\title{String search func facs}
\usage{
whole_str(haystack, needle)

part_str(haystack, needle)

both_str(string_vec)
}
\arguments{
\item{haystack}{objedt in which to look for the needle}

\item{needle}{vector of strings to look for}

\item{string_vec}{vector of two strings}
}
\description{
internal function factory for functions
\strong{whole_str} matches "whole_str string"
with any non character as boundaries accepts string as regular expression.

\strong{part_str} convenience solution to match all given strings

both_str looks for strings that contain both the given strings
}
\seealso{
Other string matching functions: 
\code{\link{getElem}},
\code{\link{recodeye}()},
\code{\link{sort_substr}()}
}
\concept{string matching functions}
\keyword{internal}
