% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va}
\alias{va}
\alias{mean}
\title{va}
\usage{
mean(x, ...)
}
\description{
Conversion of va notations
}
\details{
formula from (gregori)
Although a mean value
expressed as logMAR units is sensibly back-converted to
the Snellen fraction, the SD cannot be easily backconverted
1/200 considered CF. From 20/800 no linear relation to ETDRS.
snellen 2/200 suggest to be equivalent to 2 letters,
and 1/200 to 0 letters
logMAR <- round(-1 * log10(dec_sn), 2)
ETDRS <- round(85 + 50 * log10(dec_sn), 0)
}
\examples{
eyes(amd)
}
\seealso{
Other ophthalmic functions: 
\code{\link{eyes}()}
}
\concept{ophthalmic functions}
