% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{va_chart}
\alias{va_chart}
\title{Visual acuity conversion chart}
\format{
A data frame with 29 rows and 5 variables:
\describe{
\item{snellen_ft}{snellen VA in feet}
\item{snellen_m}{snellen VA in meter}
\item{snellen_dec}{decimal snellen VA}
\item{logMAR}{logMAR VA}
\item{ETDRS}{VA in ETDRS letters}
}
}
\usage{
data("va_chart")
}
\description{
Conversion between snellen, logMAR and ETDRS.
Snellen feet, meter and decimal supported. Three qualitative common
vision measures included (light perception, hand movement and counting fingers)
}
\details{
\itemize{
\item This chart and VA conversion formulas are based on \link{holladay},
\link{beck} and \link{gregori}. Categories \strong{(no) light perception},
\strong{counting fingers} and \strong{hand movements} are converted following \link{bach}
}
}
\keyword{datasets}
