% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cleaner.R
\name{va_cleaner}
\alias{va_cleaner}
\alias{convert_na_va}
\alias{convert_NLP}
\alias{remove_plus}
\alias{clean_va}
\alias{isNAstring}
\title{convert_na_va}
\usage{
convert_na_va(x)

convert_NLP(x, replace_PL = c(PL = "LP", NPL = "NLP"))

remove_plus(x)

clean_va(x)

isNAstring(
  x,
  full = c("\\\\.+", "", "\\\\s+", "n/a", "na", "null"),
  partial = c("not")
)
}
\arguments{
\item{x}{vector}
}
\description{
convert_na_va

convert_NLP

remove_plus

clean_va

isNAstring
}
\concept{va cleaner functions}
