% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va_dissect}
\alias{va_dissect}
\title{Converting each VA element}
\usage{
va_dissect(x)
}
\arguments{
\item{x}{vector of VA.}
}
\description{
Helper for \code{\link[=va]{va()}} if  \code{\link[=which_va]{which_va()}} finds "mixed" VA
Does require the VA vector to be prepared with \code{\link[=clean_va]{clean_va()}} first.
convert seach VA vector element individually
}
\seealso{
Other VA converter: 
\code{\link{va_methods}},
\code{\link{va}()},
\code{\link{which_va}()}
}
\concept{VA converter}
