% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_generics.R
\name{va_generics}
\alias{va_generics}
\alias{va_logmar}
\alias{va_etdrs}
\alias{va_snellen}
\title{VA conversion generics}
\usage{
va_logmar(x, ...)

va_etdrs(x, ...)

va_snellen(x, snellen = "ft")
}
\arguments{
\item{x}{vector of visual acuities}

\item{snellen}{which Snellen notation to display. Either "sn", "m" or "dec"}
}
\description{
S3 generics for VA conversion
}
\seealso{
Other VA methods: 
\code{\link{va_methods}}
}
\concept{VA generics}
\concept{VA methods}
