% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va_helper}
\alias{va_helper}
\alias{which_va}
\alias{va_dissect_apply}
\title{VA helper}
\usage{
which_va(x)

va_dissect_apply(elem)
}
\arguments{
\item{x}{Vector with VA entries}
}
\description{
Internal functions which help with VA conversion
\itemize{
\item \code{which_va}: guessing which VA notation (VA class)
}

\itemize{
\item \code{va_dissect_apply}: converts every element individually
}
}
\seealso{
Other VA converter: 
\code{\link{va}()}
}
\concept{VA converter}
\concept{VA helper}
