% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va_methods}
\alias{va_methods}
\alias{va_logmar}
\alias{va_etdrs}
\alias{va_snellen}
\alias{tologMAR.snellen}
\alias{toETDRS.snellen}
\alias{toSnellen.logmar}
\alias{toETDRS.logMAR}
\alias{tologMAR.etdrs}
\alias{toSnellen.etdrs}
\title{VA conversion}
\usage{
va_logmar(x)

va_etdrs(x)

va_snellen(x, snellen = "ft")

tologMAR.snellen(x)

toETDRS.snellen(x)

toSnellen.logmar(x, snellen = "ft")

toETDRS.logMAR(x)

tologMAR.etdrs(x)

toSnellen.etdrs(x, snellen = "ft")
}
\arguments{
\item{x}{vector of class ETDRS}

\item{snellen}{which Snellen notation to display. Either "sn", "m" or "dec"}
}
\description{
S3 methods for VA conversion
}
\details{
Only approximate eqivalent.
Using formula suggested by Gregori et al.

Only approximate eqivalent.
Rounding logMAR to nearest first digit and converting
to snellen using the va conversion chart

Only approximate eqivalent.
Rounding logMAR to nearest first digit and
converting to ETDRS letters using the va conversion chart

Approximate eqivalent based on formula in Beck et al.

Approximate eqivalent based on formula in Beck et al.
First coverting to logMAR, rounding this to the closest first digit,
converting this to snellen using the chart.
}
\concept{va conversion methods}
