% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAclassConversion.R
\name{VAclassConversion}
\alias{VAclassConversion}
\alias{to_logmar}
\alias{to_etdrs}
\alias{to_snellen}
\title{VA class conversion}
\usage{
to_logmar(x, ...)

to_etdrs(x, ...)

to_snellen(x, ...)
}
\arguments{
\item{x}{vector of visual acuities}

\item{...}{parameters passed to \link{va}}
}
\value{
vector with visual acuity of class to_(...)
See also \link{convertVA}: "VA classes"
}
\description{
Simple convenience wrapper around \link{va} to get desired VA class
}
\seealso{
Other VA converter: 
\code{\link{snellen_steps}},
\code{\link{va_dissect}()},
\code{\link{va_methods}},
\code{\link{va}()},
\code{\link{which_va}()}
}
\concept{VA converter}
