% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\docType{data}
\name{VAclasses}
\alias{VAclasses}
\title{VA classes}
\format{
An object of class \code{character} of length 5.
}
\usage{
VAclasses
}
\description{
eye introduces classes for VA notation.
If the class is not manually attributed to the vectors (not advised!), they
will have (at least) three classes:
\enumerate{
\item class "va"
\item One of class 'snellen', 'snellendec', 'logmar', 'etdrs' or 'quali'
\item Either of 'character' (for Snellen, snellendec, and qualitative),
'numeric' (for logMAR), or 'integer' (for ETDRS).
}
}
\keyword{internal}
