% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_va.R
\name{which_va}
\alias{which_va}
\alias{which_va_dissect}
\title{Guessing the VA class}
\usage{
which_va(x)

which_va_dissect(elem)
}
\arguments{
\item{x}{Vector with VA entries}

\item{elem}{element of vector to convert}
}
\description{
Guessing the VA notation (VA class)
\itemize{
\item \code{which_va}: guessing VA class for entire vector
}

\itemize{
\item \code{which_va_dissect}: guessing VA class for each vector element
}
}
\seealso{
Other VA converter: 
\code{\link{va_dissect}()},
\code{\link{va_methods}},
\code{\link{va}()}
}
\concept{VA converter}
\concept{VA helper}
