% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{age}
\alias{age}
\title{age}
\usage{
age(from_date, to_date = lubridate::now(), period = FALSE, dec = 1)
}
\arguments{
\item{from_date}{start date}

\item{to_date}{end date}

\item{period}{default FALSE: output as a duration. If TRUE, output as a period}
}
\description{
calculates age in years, either as duration or as period
Slight modification of a function \href{https://stackoverflow.com/a/47529507/7941188}{posted on stackoverflow by user Moody_Mudskipper}
}
\examples{
age("1984-10-16")

dob <-  c("1984-10-16", "2000-01-01")
test_date <-  as.Date(dob) + c(15000, 20000)
age(dob, test_date)
}
\seealso{
Other convenience functions: 
\code{\link{csv}()}
}
\concept{convenience functions}
