% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink.R
\name{blink}
\alias{blink}
\title{Your data in a blink of an eye}
\usage{
blink(x)
}
\arguments{
\item{x}{data frame}
}
\description{
High level function, wrapping \link{myop}, \link{eyes} and \link{insight}.
Checks if data is in long format. If not, myop will be applied.
eyes and insight will be applied on the result.
}
\details{
TBC
}
\examples{
TBC
}
