% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink.R
\name{blink}
\alias{blink}
\title{Your data in a blink of an eye}
\usage{
blink(x, id = NULL, eye = NULL, fct = TRUE, fct_level = 0:4, binary = TRUE)
}
\arguments{
\item{x}{data frame}

\item{id}{passed to \link{eyes}}

\item{eye}{passed to \link{eyes}}

\item{fct}{Keep columns for reveal and va that are factors}

\item{fct_level}{Remove columns for reveal and va when all unique values
fall into the range of fct_level}

\item{binary}{Remove columns for reveal and va with logicals}
}
\description{
High level function, wrapping \link{myop}, \link{eyes} and \link{reveal}
}
\details{
Steps of data processing:
\itemize{
\item checking for duplicate rows and removing them
\item Renaming columns to prepare for myopization and VA conversion
\item VA will always be converted to logmar by default
}
}
\examples{
blink(amd)

messy_df <- data.frame( id = letters[1:3],
iop_r_preop = sample(21:23), iop_r_postop = sample(11:13),
iop_l_postop = sample(11:13), iop_l_preop = sample(31:33),
va_r_preop = sample(41:43),  va_l_preop = sample(41:43),
va_r_postop = sample(51:53), va_l_postop = sample(45:47)
)
blink(messy_df)
}
