% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cleaner.R
\name{clean_va}
\alias{clean_va}
\alias{convert_NLP}
\alias{isNAstring}
\title{Visual acuity entry cleaner}
\usage{
clean_va(x)

convert_NLP(x, replace_PL = c(pl = "lp", npl = "nlp"))

isNAstring(x, full = c("\\\\.+", "", "\\\\s+", "n/a", "na", "null"))
}
\arguments{
\item{x}{Vector with VA entries}

\item{replace_PL}{named vector how to rename qualitative VA}

\item{full}{vector of full strings to be replaced by NA}
}
\description{
VA cleaning:
\enumerate{
\item \code{\link[=isNAstring]{isNAstring()}}:
Replacing empty placeholders (".","", "(any number of empty space)",
"NULL", "NA" ) with NA with
\item Removing "plus" and "minus" from snellen notation
\item \code{\link[=convert_NLP]{convert_NLP()}} Simplifying the notation for qualitative VA notation
(NPL becomes NLP, PL becomes LP) .
}
}
\seealso{
Other VA cleaner: 
\code{\link{va}()}

Other VA cleaner: 
\code{\link{va}()}
}
\concept{VA cleaner}
