% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plausible.R
\name{convertQuali}
\alias{convertQuali}
\title{convert quali entries}
\usage{
convertQuali(x, to_class)
}
\arguments{
\item{x}{vector}

\item{va_class}{to which class}
}
\value{
vector
}
\description{
convert quali entries
}
\keyword{internal}
