% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cleaner.R
\name{convertQuali}
\alias{convertQuali}
\title{convert quali entries}
\usage{
convertQuali(x, to_class)
}
\arguments{
\item{x}{vector}

\item{to_class}{to which class}
}
\value{
vector
}
\description{
converting quality VA entries
}
\keyword{internal}
