% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{count_eyes}
\alias{count_eyes}
\title{internal count}
\usage{
count_eyes(x, id_col, eye_col, details = FALSE)
}
\arguments{
\item{x}{object (data frame)}

\item{id_col}{patient column}

\item{eye_col}{eye column}

\item{details}{if TRUE, will get more information}
}
\value{
Named integer vector with count of patients and eyes
}
\description{
\code{count_eyes()} is the internal counting function for \link{eyes}
}
\keyword{internal}
