% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{csv}
\alias{csv}
\title{csv}
\usage{
csv(x, name = NULL)
}
\arguments{
\item{x}{data frame}

\item{name}{Filename (Default: Name of dataframe). If provided,
character string (.csv extension added automatically)}
}
\value{
No return value, called for side effects (saving file)
}
\description{
wrapper around write.csv with default 'row.names = FALSE'.
Will use the name of the data frame for the generated .csv file.
}
\examples{
\dontrun{
csv(amd)
}
}
\seealso{
Other convenience functions: 
\code{\link{age}()}
}
\concept{convenience functions}
