% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eye.R
\docType{package}
\name{eye}
\alias{eye}
\title{eye}
\description{
See more with \emph{eye}
}
\section{Ophthalmology functions}{

\emph{eye} is dedicated to facilitate very common tasks
in ophthalmic research.
\itemize{
\item Visual acuity conversion for snellen, logMAR and ETDRS \code{\link{va}}
\item Counting patients and eyes \code{\link{eyes}}
\item Recode eye strings \code{\link{recodeye}}
\item Reshape eye specific variables \code{\link{myop}} and \code{\link{hyperop}}
\item Summarizing data with common statistics (mean, sd, n, range)\code{\link{reveal}}
\item Easy summary of your eye data in one blink \code{\link{blink}}
}

\emph{eye} contains a real life data set \code{\link[eye:amd]{eye::amd}}
}

\section{Beyond ophthalmology}{

\itemize{
\item \code{\link{age}}: Calculate age
}
}

\seealso{
Useful links:
\itemize{
\item \href{https://github.com/tjebo/eye}{\code{eye} on github}
\item \href{https://github.com/tjebo/eye/issues}{Report bugs}
}
}
\author{
Tjebo Heeren
}
\keyword{internal}
