% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{eyes}
\alias{eyes}
\alias{counteyes}
\alias{count_eyes}
\title{eyes}
\usage{
eyes(x, id = NULL, eye = NULL, eye_str = NULL)

counteyes(x, id = NULL, eye = NULL, eye_str = NULL)

count_eyes(x, id = NULL, eye = NULL, eye_str = NULL)
}
\arguments{
\item{id}{Use if \code{eyes} does not find an id / patient column. See details id column.}

\item{eye}{Use if \code{eyes} does not find an eye colum. See details}

\item{eye_str}{Use not recommended - passed to name searching functions (details)}

\item{\strong{x}}{required. (data.frame)}
}
\description{
Looks for columns that identify patients and eyes and counts number of patients and eyes.
}
\details{
\itemize{
\item \strong{id} specify if \code{eyes} fails to identify the patient columns .
\code{eyes} searches for column names that contain the string "id" or "pat"
(any letter cases). Column names that contain both strings are prioritised.
If more than one column is found, the function will return an error.
This could cause potential confusion and then the argument identifies which column to pick.
\item \strong{eye} \code{eyes} is looking for column names that contain the string "eye" (any letter cases).
This could cause potential confusion and then the argument identifies which column to pick.
\item \strong{eye_str}: for strings coding eyes: named list of character vectors
coding right and left eyes. names need to be "r" and "l".
Default \code{list(r = c("r", "re", "od", "right"),  l = c("l", "le", "os", "left"))}
}
}
\examples{
eyes(amd)
}
\seealso{
Other ophthalmic functions: 
\code{\link{va}()}
}
\concept{ophthalmic functions}
