% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{eyes}
\alias{eyes}
\alias{counteyes}
\alias{count_eyes}
\title{eyes}
\usage{
eyes(x, id = NULL, eye = NULL)

counteyes(x, id = NULL, eye = NULL)

count_eyes(x, id = NULL, eye = NULL)
}
\arguments{
\item{x}{dataframe where patient / eye information is included.}

\item{id}{id column. If not specified, automatically selected. In this case, the patient ID column name needs to contain the strings "pat" OR "id" ignore.case = TRUE.}

\item{eye}{eye colum. If not specified, automatically selected, in which case the eye column name needs to contain the string "eye" ignore.case = TRUE}

\item{text}{default = FALSE (TRUE will return text which can be pasted for example into a markdown document). As default a named vector will be returned}
}
\description{
Looks for columns that identify patients and eyes and counts number of patients and eyes.
}
\examples{
eyes(amd)
}
\seealso{
Other ophthalmic functions: 
\code{\link{va}}
}
\concept{ophthalmic functions}
