% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{eyes}
\alias{eyes}
\alias{counteyes}
\alias{count_eyes}
\title{eyes}
\usage{
eyes(data, id = NULL, eye = NULL, text = FALSE)

counteyes(data, id = NULL, eye = NULL, text = FALSE)

count_eyes(data, id = NULL, eye = NULL, text = FALSE)
}
\arguments{
\item{id}{id column. If not specified, automatically selected. In this case, the patient ID column name needs to contain the strings "pat" OR "id" ignore.case = TRUE.}

\item{eye}{eye colum. If not specified, automatically selected, in which case the eye column name needs to contain the string "eye" ignore.case = TRUE}

\item{text}{default = FALSE (TRUE will return text which can be pasted for example into a markdown document). As default a named vector will be returned}

\item{date}{dataframe where patient / eye information is included.}
}
\description{
Looks for columns that identify patients and eyes and counts number of patients and eyes.
}
\examples{
set.seed(42)
foo <- data.frame(id = c(letters[sample(10)], letters[sample(10)] ), eyes = c("r","l"))
eyes(foo)

set.seed(42)
foo1 <- data.frame(id = c(letters[sample(10)], letters[sample(10)] ), eyes = c(NA,"l"))
set.seed(42)
foo2<- data.frame(id = c(letters[sample(10)], letters[sample(10)] ), eyes = c("r","l"))
set.seed(42)
foo3 <- data.frame(id = c(letters[sample(10)], letters[sample(10)] ), eyes = c("e","l"))
set.seed(42)
foo4 <- data.frame(id = c(letters[sample(10)], letters[sample(10)] ), eyes = c("od","le"))
foo5 <- data.frame(patient= c(letters[sample(10)], letters[sample(10)] ), eyes = c("od","le"))
foo6 <- data.frame(patient= c(letters[sample(10)], letters[sample(10)] ), eyes = c("od","le", "re", "le"))
foo7 <- data.frame(patient= c(letters[sample(10)], letters[sample(10)] ), eyes = c("od","le"), patience = 'c')
foo8 <- data.frame(patient= c(letters[sample(10)], letters[sample(10)] ), eyes = c("od","le", "re", "os"))
map(map(paste0("foo", 1:8), get), function(x) try(eyes(x)))
}
