% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{eyes_to_string}
\alias{eyes_to_string}
\title{Eye count to strings}
\usage{
eyes_to_string(x, english = "small", caps = FALSE)
}
\arguments{
\item{x}{vector of one or two}

\item{english}{Which numbers to be written in plain english:
choose "small" for numbers till 12, "all" (all numbers),
or "none" (or any other string!) for none}

\item{caps}{if TRUE, first number will have capital first letter}
}
\value{
Character string - can be directly pasted into reports
}
\description{
Eye count to strings
}
\keyword{internal}
