% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{eyes_to_string}
\alias{eyes_to_string}
\title{Eye count to strings}
\usage{
eyes_to_string(x, small_num = TRUE, para = FALSE, UK = FALSE)
}
\arguments{
\item{x}{vector of one or two}

\item{small_num}{If TRUE: writing numbers <= 12 as words}

\item{para}{If TRUE: Adding "A total of" to comply with most
journal standards and to avoid awkward long numbers.}

\item{UK}{Logical, Use UK (English) style (TRUE) or
USA (American) style (FALSE).}
}
\value{
Character string - can be directly pasted into reports
}
\description{
Eye count to strings
}
\keyword{internal}
