% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElem.R
\name{getElem}
\alias{getElem}
\alias{getElem_id}
\alias{getElem_eye}
\alias{getElem_va}
\alias{getElem_iop}
\title{Find element based or strings}
\usage{
getElem_id(obj, id_chr)

getElem_eye(obj, eye_chr)

getElem_va(obj, va_chr)

getElem_iop(obj, iop_chr)
}
\arguments{
\item{obj}{can be vector, data frame or list}

\item{id_chr}{strings to identify ID colums.}

\item{eye_chr}{named list with strings for "r" and "l"
names have to be "r" and "l"!}

\item{va_chr}{named list with strings for "whole" and "part"
names have to be "whole" and "part"!}

\item{iop_chr}{named list with strings for "whole" and "part"
names have to be "whole" and "part"!}
}
\value{
vector
}
\description{
Finds the element in an object for one of the following:
\itemize{
\item ID
\item eye
\item Visual acuity
}

\code{getElem} search a vector, the column names (data frame) or the names (list)
}
\seealso{
Other string matching functions: 
\code{\link{recodeye}()},
\code{\link{set_codes}()},
\code{\link{sort_substr}()},
\code{\link{str_func_facs}}
}
\concept{string matching functions}
\keyword{internal}
