% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElem.R
\name{getElem}
\alias{getElem}
\alias{getElem_id}
\alias{getElem_eyecol}
\alias{getElem_eye}
\alias{getElem_va}
\alias{getElem_iop}
\title{Find element based or strings}
\usage{
getElem_id(obj)

getElem_eyecol(obj)

getElem_eye(obj)

getElem_va(obj)

getElem_iop(obj)
}
\arguments{
\item{obj}{can be vector, data frame or list}
}
\value{
vector
}
\description{
Finds the element in an object for one of the following:
\itemize{
\item \emph{getElem_id} ID column
\item \emph{getElem_eyecol} eye column
\item \emph{getElem_eye} columns that contain information for right/left eyes
\item \emph{getElem_va} Visual acuity
\item \emph{getElem_iop} IOP
}

\code{getElem} search a vector, the column names (data frame) or the names (list)
}
\seealso{
Other string matching functions: 
\code{\link{recodeye}()},
\code{\link{sort_substr}()},
\code{\link{str_search}}
}
\concept{string matching functions}
\keyword{internal}
