% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fct.R
\name{whole}
\alias{whole}
\alias{partial}
\alias{get_eyecols}
\alias{get_va_cols}
\alias{get_iop_cols}
\title{whole}
\usage{
whole(string)

partial(string)

get_eyecols(x)

get_va_cols(eyecols)

get_iop_cols(eyecols)
}
\description{
internal unction factory for function
to match "whole string" with any non character as boundaries
accepts string as regular expression.

internal unction factory for function to match string
accepts string as regular expression.

internal function gets the eye columns

internal function gets the va columns

internal function gets the va columns
}
\examples{
whole("eye")(amd)

partial("pat|Id")(amd)
}
\concept{string matching functions}
