% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fct.R, R/va.R
\name{whole}
\alias{whole}
\alias{part_str}
\alias{set_eye}
\alias{set_quali}
\alias{set_iop}
\alias{set_va}
\alias{get_idcols}
\alias{get_eyecols}
\alias{get_iop_cols}
\alias{get_va_cols}
\alias{which_va}
\title{whole}
\usage{
whole(string)

part_str(string)

set_eye(r = c("r", "re", "od", "right"), l = c("l", "le", "os", "left"))

set_quali(quali = c("NLP", "LP", "HM", "CF"))

set_iop(whole = c("iop", "gat", "nct"), partial = "pressure")

set_va(whole = c("va", "bcva"), partial = "acuit")

get_idcols(obj, id_str = c("id", "pat"))

get_eyecols(obj, eye_chr)

get_iop_cols(obj, iop_chr)

get_va_cols(obj, va_chr)

which_va(x)
}
\arguments{
\item{obj}{can be vector, data frame or list}

\item{id_str}{strings to identify id colums. Default c("id", "pat")
for convenience \link{set_eye} function can and should be used.}

\item{eye_chr}{named list with strings for "r" and "l"
names have to be "r" and "l"!
for convenience \link{set_eye} function can and should be used.}

\item{iop_chr}{named list with strings for "whole" and "partial"
names have to be "whole" and "partial"!
for convenience \link{set_iop} function can and should be used.}

\item{va_chr}{named list with strings for "whole" and "partial"
names have to be "whole" and "partial"!
for convenience \link{set_va} function can and should be used.}

\item{x}{vector}
}
\description{
internal unction factory for function
to match "whole string" with any non character as boundaries
accepts string as regular expression.

internal unction factory for function to match string
accepts string as regular expression.

set the eye strings looked for in the data

set the iop strings looked for in the data

set the va strings looked for in the data

internal function gets the eye columns

internal function gets the eye columns

internal function gets the iop columns

internal function gets the va columns

internal check which type of va in order to assign class.
}
\examples{
whole("eye")(amd)

part_str("pat|Id")(amd)
}
\concept{string matching functions}
\concept{va conversion functions}
