% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myop.R
\name{myop}
\alias{myop}
\alias{myope}
\alias{myopise}
\alias{myopize}
\alias{myopic}
\title{long eye data}
\usage{
myop(x, cols, id, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myope(x, cols, id, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myopise(x, cols, id, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myopize(x, cols, id, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myopic(x, cols, id, values_to, names_to = "eye", eye_code = c("r", "l"), ...)
}
\arguments{
\item{x}{data frame}

\item{cols}{Eye columns either automatic or with vector of two - see details}

\item{id}{id column to gather by - relevant if gathering multiple variables}

\item{values_to}{passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}

\item{names_to}{passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}

\item{eye_code}{Vector of two with right eye coded first.}

\item{...}{other arguments passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}
}
\description{
Wrapper around \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} to make your data
long ("myopic")
}
\details{
\itemize{
\item \strong{cols}: if not specified, \code{myop()} will look for columns that have
typical eye codes, i.e. c("r", "re" and "od") for right eyes and
c("l", "le" and "os") for left eyes.
\item if argument specified, right eye has to be given first!
\item \strong{eye_code}: numeric codes of 0:1 or 1:2 are allowed, other
numeric codes are not supported.
}
}
\examples{
set.seed(42)
iop <- data.frame(id = letters[1:11], r = sample(10:20), l = sample(10:20))

myop(iop, values_to = "iop")

# Example to clean a bit messy data frame
iop_va <- data.frame(id = letters[1:11],
                     iop_r = sample(10:20), iop_l = sample(10:20),
                     va_r = sample(40:50), va_l = sample(40:50))
myop(iop_va)

}
