% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyes.R
\name{myop}
\alias{myop}
\alias{myope}
\alias{myopise}
\alias{myopize}
\title{long eye data}
\usage{
myop(x, cols, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myope(x, cols, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myopise(x, cols, values_to, names_to = "eye", eye_code = c("r", "l"), ...)

myopize(x, cols, values_to, names_to = "eye", eye_code = c("r", "l"), ...)
}
\arguments{
\item{x}{data frame}

\item{cols}{Eye columns either automatic or with vector of two - see details}

\item{values_to}{passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}

\item{names_to}{passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}

\item{eye_code}{Vector of two with right eye coded first.}

\item{...}{other arguments passed to \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}}
}
\description{
Wrapper around \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} to make your data
long ("myope")
}
\details{
\itemize{
\item \strong{cols}: if not specified, \code{myop()} will look for columns that have
typical eye codes, i.e. c("r", "re" and "od") for right eyes and
c("l", "le" and "os") for left eyes.
\item if argument specified, right eye has to be given first!
\item \strong{eye_code}: numeric codes of 0:1 or 1:2 are allowed, other
numeric codes are not supported.
}
}
\examples{
set.seed(42)
iop <- data.frame(id = letters[1:11], r = sample(10:20), l = sample(10:20))

myop(iop, values_to = "iop")

# Example to clean a messy data frame
iop_va <- data.frame(id = letters[1:11],
                     iop_r = sample(10:20), iop_l = sample(10:20),
                     va_r = sample(40:50), va_l = sample(40:50))
head(iop_va)
# use myope twice on both iop and va columns
iop_long <- myop(iop_va, cols = c("iop_r", "iop_l"), values_to = "iop")
va_long <- myop(iop_va, cols = c("va_r", "va_l"), values_to = "va")
# full join both data frames
iop_va_clean <- full_join(iop_long, va_long, by = c("id", "eye")) \%>\%
  select(id, eye, va, iop)
head(iop_va_clean)

}
