% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myop.R
\name{myop_rename}
\alias{myop_rename}
\title{Rename}
\usage{
myop_rename(x)
}
\arguments{
\item{x}{object (data frame)}
}
\description{
Rename data names for \code{\link[=myop]{myop()}}, replacing "." with "_"
in order to prepare for the use of \code{\link[=sort_substr]{sort_substr()}} - sorting
eye strings first, then strings coding for methods like IOP/VA,
then the rest.
}
\seealso{
Other myopizer: 
\code{\link{myop_pivot}()},
\code{\link{myopizer}()}
}
\concept{myopizer}
\keyword{internal}
