% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_methods}
\alias{print_methods}
\alias{print.snellen}
\alias{print.logmar}
\alias{print.etdrs}
\alias{print.blink}
\title{print eye classes}
\usage{
\method{print}{snellen}(x, ...)

\method{print}{logmar}(x, ...)

\method{print}{etdrs}(x, ...)

\method{print}{blink}(x, ...)
}
\arguments{
\item{x}{object of class "blink"}

\item{...}{arguments passed to \link{print.default}}
}
\value{
No return value, called for side effects (printing)
}
\description{
S3 methods for VA classes "snellen", "logmar" and "etdrs".
\strong{snellen} is always also a character class- because it is more categorical
than continuous. \strong{logmar} and \strong{etdrs} are both numerics
(logMAR is double, etdrs is integer).

S3 methods for blink class
}
