% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print_methods}
\alias{print_methods}
\alias{print.snellen}
\alias{print.logmar}
\alias{print.etdrs}
\alias{print.blink}
\alias{print.eyes}
\alias{print.eyes_details}
\title{print eye classes}
\usage{
\method{print}{snellen}(x, ...)

\method{print}{logmar}(x, ...)

\method{print}{etdrs}(x, ...)

\method{print}{blink}(x, ...)

\method{print}{eyes}(x, ...)

\method{print}{eyes_details}(x, show = 6, ...)
}
\arguments{
\item{x}{object of class "eyes_details"}

\item{...}{arguments passed to \link{print.default}}

\item{show}{how many subjects to be shown before printing the footnote}
}
\value{
No return value, called for side effects (printing)
}
\description{
S3 methods for VA classes "snellen", "logmar" and "etdrs".
\strong{snellen} is always also a character class- because it is more categorical
than continuous. \strong{logmar} and \strong{etdrs} are both numerics
(logMAR is double, etdrs is integer).

S3 methods for class blink

S3 methods for class eyes

S3 methods for class eyes_details
}
