% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink.R
\name{remCols}
\alias{remCols}
\title{Remove cols from selected cols}
\usage{
remCols(x, cols, fct, fct_level, binary)
}
\arguments{
\item{x}{data frame}

\item{cols}{cols}

\item{fct}{Keep columns for reveal and va that are factors}

\item{fct_level}{Remove columns for reveal and va when all unique values
fall into the range of fct_level}

\item{binary}{Remove columns for reveal and va with logicals}
}
\description{
avoiding colums that only
contain values fct_levels or binary,
because they are likely categorical codes.
}
