% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink.R
\name{remCols}
\alias{remCols}
\title{Remove cols from selected cols}
\usage{
remCols(x, cols, fct_level)
}
\arguments{
\item{x}{data frame}

\item{cols}{cols}

\item{fct_level}{Remove columns for reveal and va when all unique values
fall into the range of fct_level}
}
\value{
logical vector
}
\description{
Helper for \code{\link[=blink]{blink()}}. Removes columns that only
contain values fct_levels or binary from selected columns (currently for
both automatically and manually selected columns).
fct_levels are removed because they are likely categorical codes.
}
\keyword{internal}
