% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reveal}
\alias{reveal}
\title{reveal}
\usage{
reveal(x, dec = 1, dataframe = TRUE)
}
\arguments{
\item{x}{either vector or list of vectors. Data frames supported.
Character vectors (or character columns) will be removed.}

\item{dec}{how many decimals are displayed}

\item{dataframe}{returning list if FALSE}
}
\value{
data frame (default) or list
}
\description{
Shows commonly used summary statistics
}
\examples{
x = y = z = c(rnorm(20), NA)

# named or unnamed list
mylist <- list(x = x, y = y, z = z)
reveal(mylist)
# with a data frame
mydf <- data.frame(x, y, z)
reveal(mydf)
#If aggregation by group, split the data frame first
mydf2 <- data.frame(group = rep(letters[1:2], each = 42), x, y, z)
lapply(split(mydf2, mydf2$group), see, rownames = FALSE)
}
\concept{stats functions}
