% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal.R
\name{reveal_methods}
\alias{reveal_methods}
\alias{revealEye}
\alias{revealEye.list}
\alias{revealEye.numeric}
\alias{revealEye.data.frame}
\alias{revealEye.default}
\title{reveals little helper}
\usage{
revealEye(x, ...)

\method{revealEye}{list}(x, by, dec, funs, ...)

\method{revealEye}{numeric}(x, dec, funs, ...)

\method{revealEye}{data.frame}(x, dec, funs, ...)

\method{revealEye}{default}(x, dec, funs, ...)
}
\arguments{
\item{x}{data frame, numeric vector, or list of numeric vectors}

\item{...}{further arguments passed to methods}

\item{by}{character vector with the names of the columns. Can be
several variables!}

\item{dec}{how many decimals are displayed}

\item{funs}{not really meant to be used at the moment - change the
Summarizing functions with a named(!) list of functions}
}
\value{
data frame
}
\description{
S3 generic and methods
}
\seealso{
Other revealer: 
\code{\link{reveal}()},
\code{\link{reveal_split}()}
}
\concept{revealer}
\keyword{internal}
