% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_methods.R
\name{snellen_steps}
\alias{snellen_steps}
\alias{snellensteps}
\title{Convert plus minus entries}
\usage{
snellensteps(x, smallstep)
}
\arguments{
\item{x}{Vector with VA entries of class snellen - needs to be in
format xx/yy}

\item{smallstep}{if plusminus shall be considered as logmar equivalent}
}
\value{
character vector of Snellen entries
}
\description{
Convert plus minus entries
}
\seealso{
https://en.wikipedia.org/wiki/Psychometric_function

Other VA converter: 
\code{\link{VAwrapper}},
\code{\link{plausibility_methods}},
\code{\link{va_methods}},
\code{\link{va_mixed}()},
\code{\link{va}()},
\code{\link{which_va}()}
}
\concept{VA converter}
\keyword{internals}
