% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_helper.R
\name{str_search}
\alias{str_search}
\alias{whole_str}
\alias{part_str}
\alias{both_str}
\title{String search helper}
\usage{
whole_str(haystack, needle)

part_str(haystack, needle)

both_str(haystack, needle)
}
\arguments{
\item{haystack}{object in which to look for the needle}

\item{needle}{vector of strings to look for - accepts regular expressions}
}
\description{
\strong{whole_str} finds haystack with "whole needles"
with any non character as boundaries

\strong{part_str} finds haystacks that match any of the needles

\strong{both_str} will find haystacks that contain both needles
}
\seealso{
Other string matching functions: 
\code{\link{getElem}},
\code{\link{recodeye}()},
\code{\link{sort_substr}()}
}
\concept{string matching functions}
\keyword{internal}
