% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va}
\alias{va}
\title{va}
\usage{
va(x, from = NULL, to = "logmar")
}
\arguments{
\item{x}{vector with VA values}

\item{from}{String, from which notation to convert (See details)}

\item{to}{String, to which notation to convert. Default logMAR (See details)}
}
\description{
VA notation conversion
}
\details{
\strong{from} if not provided (default), \code{va} will try to guess.
This will work in most cases, but can fail due to weird, unpredicted ways
to enter and save data. For more control, provide string:
Either "ETDRS", "logMAR" or "snellen" (any case allowed).
\strong{to} Supported: "ETDRS", "logMAR" or "snellen" (any case allowed).

Although a mean value
expressed as logMAR units is sensibly back-converted to
the Snellen fraction, the SD cannot be easily backconverted
1/200 considered CF. From 20/800 downwards no linear relation to ETDRS.
snellen 2/200 suggest to be equivalent to 2 letters,
and 1/200 to 0 letters
}
\examples{
TBC
}
\seealso{
Other ophthalmic functions: 
\code{\link{eyes}()}
}
\concept{ophthalmic functions}
