% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va}
\alias{va}
\alias{convert_va}
\alias{va_dissect}
\title{Visual acuity notation conversion}
\usage{
va(x, from = NULL, to = "logmar", type = "ft")

convert_va(x, from = NULL, to = "logmar", type = "ft")

va_dissect(x)
}
\arguments{
\item{x}{Vector with visual acuity entries. Must be atomic.}

\item{from}{from which class to convert. "etdrs", "logmar" or "snellen" -
any case allowed. Ignored if it does not make sense}

\item{to}{to which class to convert. "etdrs", "logmar" or "snellen" -
any case allowed}

\item{type}{if to = "snellen", which snellen notation.
"m", "dec" or "ft"}
}
\value{
\itemize{
\item \strong{va}: vector of class set with \code{to} argument
\item \strong{va_dissect}: Named vector of class \code{logmar}
}
}
\description{
Cleans VA and converts VA notations (classes) into one another.
\code{va} will guess the VA class based on specific rules (see details)
}
\details{
VA can be snellen feet/meter/decimal, logMAR, ETDRS, or
"qualitative" (Couting fingers, etc.), and a mix of those!
\itemize{
\item Snellen fractions need to be either form 6/x or 20/x
\item ETDRS must be between 0 and 100
\item logMAR must be between -0.3 and 3.0
\item Qualitative must be PL, LP, NLP, NPL, HM, CF (any case allowed)
For further conversion rules see section "VA guessing"
}

\strong{\code{va}} cleans visual acuity entries with \link{clean_va}: it assigns \code{NA}
to missing entries, removes "plus" and "minus" from snellen entries
and unifies the notation for qualitative entries.

\strong{\code{va_dissect}} is very similar to \code{va}. However, it is less flexible
and will always convert to logmar. It also does not support
snellen decimals! Usually you don't need to use it -
It is internally called when \link{which_va} guesses "mixed" notation.

Distinction between logMAR and snellen decimals, or between ETDRS and
logMAR can be ambiguous in some special cases and \strong{logmar} is picked -
provide \emph{from} if you want something else.

\strong{snellen decimals} are a particular challenge and \code{va} may wrongly
assign  logMAR - this could happen if you have weird snellen decimal
values that are not found in \link{va_chart}. E.g., check with \code{unique(x)}.

\emph{quali} class represents "(No) light perception", "hand movements"
and "counting fingers" and are converted to logmar following the
suggestions by \href{https://michaelbach.de/sci/acuity.html}{Michael Bach}
}
\section{VA guessing}{

\itemize{
\item if x integer and 3 < x <= 100: \code{etdrs}
\item if x integer and 0 <= x <= 3: \code{logmar}, but you can choose \code{etdrs}
\item if x numeric and -0.3 <= x <= 3: \code{logmar}
\item if x numeric and all x in intersection(va_chart$logMAR, va_chart$snellen_dec):
\code{logmar}, but you can choose \code{snellen}
\item \emph{non-mixed class}: if all x in va_chart$snellen_dec: \code{snellen}
\item \emph{mixed class} (\link{which_va_dissect}): snellen_dec not supported.
\item if character and format x/y: \code{snellen} (fraction)
\item if one of "CF", "HM", "LP", "PL", "NLP", or "NPL": \code{quali}
\item if numeric x beyond the ranges from above: \code{NA}
\item Any other string or NA: \code{NA}
}
}

\examples{
## will automatically detect VA class and convert to logMAR by default
## ETDRS letters
x <- c(23, 56, 74, 58)
va(x)

## ... or convert to snellen
va(x, to = "snellen")

## snellen, mixed with categories. Also dealing with those "plus/minus" entries
va(c("NLP", "NPL", "PL", "LP", "HM", "CF", "6/60", "20/200", "6/9",
 "20/40", "20/40+3", "20/50-2"))

}
\seealso{
Other VA converter: 
\code{\link{va_methods}},
\code{\link{which_va}()}

Other VA cleaner: 
\code{\link{va_cleaner}}
}
\concept{Ophthalmic functions}
\concept{VA cleaner}
\concept{VA converter}
