% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va}
\alias{va}
\title{va}
\usage{
va(x, from = NULL, to = "logmar", snellen_type = "ft")
}
\arguments{
\item{x}{vector with VA values}

\item{from}{String, from which notation to convert (See details)}

\item{to}{String, to which notation to convert. Default logMAR (See details)}

\item{snellen_type}{which snellen method to use when \emph{to = "snellen"}.
Either "m", "dec" or "ft"}
}
\description{
VA notation conversion
}
\details{
\strong{from} if not provided (default), \code{va} will try to guess.
This will work in most cases, but can fail due to weird, unpredicted ways
to enter and save data. For more control, provide string:
Either "ETDRS", "logMAR" or "snellen" (any case allowed).
\strong{to} Supported: "ETDRS", "logMAR" or "snellen" (any case allowed).

Qualitative vision measures "Light perception", "hand movements" and "counting fingers"
have to be coded "LP", "HM" and "CF", otherwise they will be replaced with NA.
"No light perception" is not converted into visual acuity values and will be replaced with NA.
}
\examples{
TBC
}
\seealso{
Other ophthalmic functions: 
\code{\link{eyes}()},
\code{\link{insight}()}
}
\concept{ophthalmic functions}
