% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_cleaner.R
\name{va_cleaner}
\alias{va_cleaner}
\alias{clean_va}
\title{Visual acuity entry cleaner}
\usage{
clean_va(x)
}
\arguments{
\item{x}{Vector with VA entries}
}
\description{
Internal functions for VA cleaning.
\code{va_cleaner}:
\itemize{
\item \code{\link[=isNAstring]{isNAstring()}}:
Replacing empty placeholders (".","", "(any number of empty space)",
"NULL", "NA" ) with NA with
\item Removing "plus" and "minus" from snellen notation
\item \code{\link[=convert_NLP]{convert_NLP()}} Simplifying the notation for qualitative VA notation
(NPL becomes NLP, PL becomes LP) .
}
}
\seealso{
Other VA cleaner: 
\code{\link{convert_NLP}()},
\code{\link{va}()}
}
\concept{VA cleaner}
