% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va.R
\name{va_dissect}
\alias{va_dissect}
\title{Converting each VA element}
\usage{
va_dissect(x, from_logmar, to, ...)
}
\arguments{
\item{x}{vector of VA.}

\item{to}{to which VA notation}

\item{...}{passed to convertVA}
}
\value{
vector of \code{va} class. see also \code{\link[=va]{va()}}
}
\description{
Used in \code{\link[=va]{va()}} if  \code{\link[=which_va]{which_va()}} finds "mixed" VA notation.
Converts each VA vector element individually - requires the VA vector to
be prepared with \code{\link[=clean_va]{clean_va()}} first.
}
\seealso{
Other VA converter: 
\code{\link{snellen_steps}},
\code{\link{va_methods}},
\code{\link{va}()},
\code{\link{which_va}()}
}
\concept{VA converter}
\keyword{internal}
