% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_methods.R
\name{va_methods}
\alias{va_methods}
\alias{convertVA}
\alias{convertVA.quali}
\alias{convertVA.snellendec}
\alias{convertVA.snellen}
\alias{convertVA.logmar}
\alias{convertVA.etdrs}
\alias{convertVA.default}
\title{VA conversion methods}
\usage{
convertVA(x, ...)

\method{convertVA}{quali}(x, to, type, ...)

\method{convertVA}{snellendec}(x, to, type, ...)

\method{convertVA}{snellen}(x, to, type, smallstep, noplus, ...)

\method{convertVA}{logmar}(x, to, type, ...)

\method{convertVA}{etdrs}(x, to, type, ...)

\method{convertVA}{default}(x, to, ...)
}
\arguments{
\item{x}{vector of visual acuities}

\item{...}{further arguments passed to methods}

\item{to}{to which VA class to convert}

\item{type}{which snellen notation. One of "ft", "m" or "dec"}

\item{smallstep}{how plus/minus entries are evaluated. Default to
increase/decrease snellen fractions by lines. If TRUE, each snellen
optotype will be considered equivalent to 0.02 logmar or 1 ETDRS
letter (assuming 5 letters in a row in a chart)}

\item{noplus}{ignoring plus/minus entries and just returning the
snellen fraction. This overrides the smallstep argument.}
}
\value{
vector with visual acuity of class \code{va}. See also "VA classes"
}
\description{
S3 methods for VA conversion
}
\details{
VA can be snellen feet/meter/decimal, logMAR, ETDRS, or
"qualitative" (Counting fingers, etc.)
\itemize{
\item Snellen fractions need to be either form 6/x or 20/x
\item ETDRS must be between 0 and 100
\item logMAR must be between -0.3 and 3.0
\item Qualitative must be PL, LP, NLP, NPL, HM, CF (any case allowed)
}

Any element which is not recognized will be converted to NA

For other conversion rules see \link{va}
}
\section{Conversion}{

Although there seems to be no good statistical reason to convert
back to Snellen, it is a very natural thing to eye specialists to think
in Snellen. A conversion to snellen gives a good gauge of how the visual
acuity for the patients are. However, back-conversion should not be
considered an exact science and any attempt to use formulas will result
in very weird Snellen values that have no correspondence to common charts.
Therefore, Snellen matching the nearest ETDRS and logMAR value in
the \link{va_chart} are used.

Further:
\itemize{
\item logMAR to ETDRS: logMAR rounded to the first digit and converted with
the chart.
\item Snellen to logMAR: logMAR = -1 * log10(snellen_frac)
\item Snellen to ETDRS: ETDRS = 85 + 50 * log10(snellen_frac)
\href{https://doi.org/10.1097/iae.0b013e3181d87e04}{Gregori et al.}.
\item ETDRS to logMAR: logMAR = -0.02 * etdrs + 1.7
\href{https://doi.org/10.1016/s0002-9394(02)01825-1}{Beck et al.}
}
}

\section{VA classes}{

convert_VA returns a vector of three classes:
\enumerate{
\item \code{va}
\item One of \code{snellen}, \code{logmar}, \code{etdrs} or \code{quali}.
\item Either of \code{character} (for Snellen and qualitative),
\code{numeric} (for logMAR), or \code{integer} (for ETDRS).
}
}

\section{plus/minus entries}{

The following rules for plus minus notations will be applied:
\itemize{
\item if entry -2 to +2 : take same Snellen value
\item if < -2 : take Snellen value one line below
\item if > +2: Snellen value one line above
}

Snellen are unfortunately often entered with "+/-", which is a
violation of a psychophysical method designed to assign one
unambiguous value to visual acuity, with
non-arbitrary thresholds based on psychometric functions. Therefore,
transforming "+/-" notation to actual results is in itself
problematic and the below suggestion to convert it will remain an
approximation to the most likely "true" result. Even more so, as the
given conditions should work for charts with
4 or 5 optotypes in a line, and visual acuity is not always tested
on such charts. Yet, I believe that the approach is still better than
just omitting the letters or (worse) assigning a missing value to those
entries.
}

\seealso{
Other VA converter: 
\code{\link{VAwrapper}},
\code{\link{plausibility_methods}},
\code{\link{snellen_steps}},
\code{\link{va}()},
\code{\link{which_va}()}
}
\concept{VA converter}
