% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_methods.R
\name{va_methods}
\alias{va_methods}
\alias{tologMAR.quali}
\alias{toSnellen.quali}
\alias{toETDRS.quali}
\alias{tologMAR.snellen}
\alias{toETDRS.snellen}
\alias{toSnellen.snellen}
\alias{toSnellen.logmar}
\alias{toETDRS.logmar}
\alias{tologMAR.logmar}
\alias{toETDRS.etdrs}
\alias{tologMAR.etdrs}
\alias{toSnellen.etdrs}
\title{VA conversion methods}
\usage{
tologMAR.quali(x)

toSnellen.quali(x, snellen = "ft")

toETDRS.quali(x)

tologMAR.snellen(x)

toETDRS.snellen(x)

toSnellen.snellen(x, snellen = "ft")

toSnellen.logmar(x, snellen = "ft")

toETDRS.logmar(x)

tologMAR.logmar(x)

toETDRS.etdrs(x)

tologMAR.etdrs(x)

toSnellen.etdrs(x, snellen = "ft")
}
\arguments{
\item{x}{Vector with VA entries}

\item{snellen}{which Snellen notation to display. Either "sn", "m" or "dec"}
}
\description{
\code{toETDRS.snellen}: Using formula suggested by Gregori et al.
ETDRS = 85 + 50 * log10(snellen_frac)

\code{toSnellen.logmar}: Rounding logMAR to first digit and
matching with Snellen entries in \link{va_chart}

\code{toETDRS.logmar}: Rounding logMAR to first digit and
matching with ETDRS letters in \link{va_chart}

\code{tologMAR.etdrs}: Using formula in Beck et al.
logMAR = -0.02 * ETDRS + 1.7

\code{toSnellen.etdrs}: Converting to logMAR with
logMAR = -0.02 * ETDRS + 1.7, rounding this to the first digit,
and matching the values to Snellen entries in \link{va_chart}
}
\seealso{
Other VA methods: 
\code{\link{va_generics}}
}
\concept{VA methods}
